modularize questions and export/import
